ktls: add boringssl support
One complexity in using kTLS is extracting the relevant session state
from the SSL library after the handshake completes: key, salt and seq.

The existing OpenSSL 1.0 variant of ktls.c manually redeclares various
structs to gain access, a fragile method and one that does not even
work for OpenSSL 1.1 (a TODO is to add OpenSSL 1.1).

Add BoringSSL support and demonstrate its interfaces to extract the
session state.

Note that BoringSSL interfaces are expressly not stable APIs, so this
is not guaranteed to continue to work in the future as is.

Signed-off-by: Willem de Bruijn <willemb@google.com>