tests: add toeplitz + rss tests
Verify

1. the rx_hash returned by a network device
2. the rss algorithm used to select an rx queue based on this rx_hash

For the first, open a PF_PACKET socket with TPACKET_V3 ring, which
is the only interface to receive the rx_hash from the kernel. Manually
compute the toeplitz hash in software and compare the two.

For the second, open one such socket for each cpu in the system, add
all into a PACKET_FANOUT_CPU group. Then compare the arrival socket,
and thus cpu, with the choice computed in software.

Signed-off-by: Willem de Bruijn <willemb@google.com>