Add option to toeplitz test to verify RPS functionality
Add the option '-r <rps_bitmap>' to toeplitz.c, which verifies that
frames are handled by CPUs according to the RPS specification.

Add the option '-rps <rps_bitmap>' to toeplitz.sh as a hook for
specifying this test. This test requires that rps_bitmap is applied
to all rx queues (i.e. all /sys/class/net/${DEV}/queues/rx-*/rps_cpus)
so that RSS is rendered effectively immaterial; if this condition is
not met, then toeplitz.sh does not proceed with the RPS test.

The test verifies that the CPU that processes each frame is chosen by
using hash % <number of "on" CPUs in the bitmap> as an index into the
bitmap, in the same way as get_rps_cpu.

Users should specify either '-C <cpu_list>' or '-r <rps_bitmap>' to
test either RSS or RPS, respectively.

Signed-off-by: Tanner Love <tannerlove@google.com>